[style] GoldPreview 를 GoldDistribution, WeeklyGoldPreview 로 분리 ( GoldDistribution : 레벨 분포 및 원형 차트)
